requirements.txt
Keras==2.4.3
matplotlib==3.3.1
mypy==0.782
mypy-extensions==0.4.3
numpy==1.18.5
pandas==1.1.2
Pillow==9.0.0
scikit-learn==1.1.1
scipy==1.4.1
seaborn==0.11.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0